fix(tests/perf): move publishRunDir + publishArtifactsIfConfigured into a non-test file
Both lived in harness_test.go but were called from harness_artifacts.go
(non-test). go test compiled fine because it includes _test.go files,
but `go build ./tests/perf/...` (or `go vet ./...`) fails with
`undefined: publishArtifactsIfConfigured` because non-test compilations
don't see _test.go declarations.

Splits them into a focused harness_publish.go file. The test seam stays
intact — harness_publisher_test.go's `publishRunDir = func(…) { … }`
still works because the var is package-level. All publisher_test cases
pass, and `go build -tags kubernetes ./tests/perf/...` is now clean.
